+++ generators/dashboard/templates/packages/components/src/Failure.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 
 export default function Failure({ error, refetch }) {
   return <div className="failure">Loading failure</div>;