+++ generators/dashboard/templates/packages/components/src/index.js
@@ -1,8 +1,8 @@
-export { Form, FormControl, FormSubmit } from '@swiftcarrot/react-form';
-export FormGroup from './FormGroup';
-export Table from './Table';
-export Sidebar from './Sidebar';
-export ErrorBoundary from './ErrorBoundary';
-export Loading from './Loading';
-export Failure from './Failure';
-export PageHeader from './PageHeader';
+export { Form, FormControl, FormSubmit } from "@swiftcarrot/react-form";
+export FormGroup from "./FormGroup";
+export Table from "./Table";
+export Sidebar from "./Sidebar";
+export ErrorBoundary from "./ErrorBoundary";
+export Loading from "./Loading";
+export Failure from "./Failure";
+export PageHeader from "./PageHeader";