+++ generators/dashboard/templates/packages/dashboard/src/layouts/Layout.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import { Sidebar } from 'components';
+import React from "react";
+import { Sidebar } from "components";
 
 export default function Layout({ children }) {
   return (