+++ generators/dashboard/templates/packages/dashboard/src/pages/graphiql.js
@@ -1,12 +1,12 @@
-import 'graphiql/graphiql.min.css';
-import React from 'react';
-import GraphiQL from 'graphiql';
-import PageHeader from 'components/src/PageHeader';
+import "graphiql/graphiql.min.css";
+import React from "react";
+import GraphiQL from "graphiql";
+import PageHeader from "components/src/PageHeader";
 
 function graphQLFetcher(graphQLParams) {
-  return fetch('http://localhost:8080/graphql', {
-    method: 'post',
-    headers: { 'Content-Type': 'application/json' },
+  return fetch("http://localhost:8080/graphql", {
+    method: "post",
+    headers: { "Content-Type": "application/json" },
     body: JSON.stringify(graphQLParams),
   }).then((response) => response.json());
 }