+++ generators/dashboard/templates/packages/dashboard/src/pages/index.js
@@ -1,2 +1,2 @@
-export HomePage from './home';
-export GraphiQLPage from './graphiql';
+export HomePage from "./home";
+export GraphiQLPage from "./graphiql";