+++ generators/packages/templates/packages/app/src/index.js
@@ -1,8 +1,8 @@
-import React from 'react';
-import { render } from 'react-dom';
+import React from "react";
+import { render } from "react-dom";
 
 const App = () => {
   return <div>App</div>;
 };
 
-render(<App />, document.getElementById('root'));
+render(<App />, document.getElementById("root"));