+++ generators/packages/templates/packages/storybook-preset/src/index.js
@@ -1,12 +1,12 @@
-const glob = require('glob');
-const rules = require('webpacker/src/webpack/rules');
+const glob = require("glob");
+const rules = require("webpacker/src/webpack/rules");
 
 const alias = glob
-  .sync('packages/**/package.json', {
-    ignore: '**/node_modules/**/package.json',
+  .sync("packages/**/package.json", {
+    ignore: "**/node_modules/**/package.json",
   })
   .reduce((x, file) => {
-    const { name } = require(file.replace('packages/', ''));
+    const { name } = require(file.replace("packages/", ""));
     x[name] = `${name}/src`;
     return x;
   }, {});