+++ generators/packages/templates/packages/webpacker/src/cmd/watch.js
@@ -1,5 +1,5 @@
-const webpack = require('webpack');
-const webpackConfig = require('../webpack');
+const webpack = require("webpack");
+const webpackConfig = require("../webpack");
 
 module.exports = function () {
   const compiler = webpack(webpackConfig);