+++ generators/packages/templates/packages/webpacker/src/utils/getStyleLoaders.js
@@ -1,40 +1,40 @@
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const postcssNormalize = require('postcss-normalize');
-const { isEnvDevelopment, isEnvProduction, paths } = require('./config');
-const shouldUseSourceMap = process.env.GENERATE_SOURCEMAP !== 'false';
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const postcssNormalize = require("postcss-normalize");
+const { isEnvDevelopment, isEnvProduction, paths } = require("./config");
+const shouldUseSourceMap = process.env.GENERATE_SOURCEMAP !== "false";
 
 //
 // https://github.com/facebook/create-react-app/blob/66bf7dfc43350249e2f09d138a20840dae8a0a4a/packages/react-scripts/config/webpack.config.js#L92
 //
 const getStyleLoaders = (cssOptions, preProcessor) => {
   const loaders = [
-    isEnvDevelopment && require.resolve('style-loader'),
+    isEnvDevelopment && require.resolve("style-loader"),
     isEnvProduction && {
       loader: MiniCssExtractPlugin.loader,
       // css is located in `static/css`, use '../../' to locate index.html folder
       // in production `paths.publicUrlOrPath` can be a relative path
-      options: paths.publicUrlOrPath.startsWith('.')
-        ? { publicPath: '../../' }
+      options: paths.publicUrlOrPath.startsWith(".")
+        ? { publicPath: "../../" }
         : {},
     },
     {
-      loader: require.resolve('css-loader'),
+      loader: require.resolve("css-loader"),
       options: cssOptions,
     },
     {
       // Options for PostCSS as we reference these options twice
       // Adds vendor prefixing based on your specified browser support in
       // package.json
-      loader: require.resolve('postcss-loader'),
+      loader: require.resolve("postcss-loader"),
       options: {
         // Necessary for external CSS imports to work
         // https://github.com/facebook/create-react-app/issues/2677
-        ident: 'postcss',
+        ident: "postcss",
         plugins: () => [
-          require('postcss-flexbugs-fixes'),
-          require('postcss-preset-env')({
+          require("postcss-flexbugs-fixes"),
+          require("postcss-preset-env")({
             autoprefixer: {
-              flexbox: 'no-2009',
+              flexbox: "no-2009",
             },
             stage: 3,
           }),
@@ -51,7 +51,7 @@ const getStyleLoaders = (cssOptions, preProcessor) => {
   if (preProcessor) {
     loaders.push(
       {
-        loader: require.resolve('resolve-url-loader'),
+        loader: require.resolve("resolve-url-loader"),
         options: {
           sourceMap: isEnvProduction ? shouldUseSourceMap : isEnvDevelopment,
           // TODO: removed for storybook-preset