+++ generators/packages/templates/packages/webpacker/src/webpack/base.js
@@ -1,13 +1,13 @@
-const webpack = require('webpack');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const WebpackAssetsManifest = require('webpack-assets-manifest');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const { paths, isEnvProduction } = require('../utils/config');
-const rules = require('./rules');
+const webpack = require("webpack");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const WebpackAssetsManifest = require("webpack-assets-manifest");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const { paths, isEnvProduction } = require("../utils/config");
+const rules = require("./rules");
 
 module.exports = {
   entry: {
-    index: './src/index.js',
+    index: "./src/index.js",
   },
   output: {
     path: paths.appBuild,
@@ -26,15 +26,15 @@ module.exports = {
 
     new MiniCssExtractPlugin({
       filename: isEnvProduction
-        ? 'static/[name].[contenthash:8].css'
-        : 'static/[name].css',
+        ? "static/[name].[contenthash:8].css"
+        : "static/[name].css",
       chunkFilename: isEnvProduction
-        ? 'static/[name].[contenthash:8].chunk.css'
-        : 'static/[name].chunk.css',
+        ? "static/[name].[contenthash:8].chunk.css"
+        : "static/[name].chunk.css",
     }),
 
     new WebpackAssetsManifest({
-      output: 'assets-manifest.json',
+      output: "assets-manifest.json",
       entrypoints: true,
       publicPath: true,
       writeToDisk: true,