+++ generators/packages/templates/packages/webpacker/src/webpack/index.js
@@ -1,5 +1,5 @@
-const { isEnvProduction } = require('../utils/config');
+const { isEnvProduction } = require("../utils/config");
 
 module.exports = isEnvProduction
-  ? require('./production')
-  : require('./development');
+  ? require("./production")
+  : require("./development");